Small refactors and bug fixes. (#2281)
