+++ python/tvm/relay/__init__.py
@@ -5,6 +5,7 @@ from ..api import register_func
 from . import base
 from . import ty
 from . import expr
+from . import expr_functor
 from . import module
 from . import ir_pass
 from .build_module import build, build_config, create_executor
@@ -52,6 +53,10 @@ Call = expr.Call
 Let = expr.Let
 If = expr.If
 TupleGetItem = expr.TupleGetItem
+
+# ExprFunctor
+ExprFunctor = expr_functor.ExprFunctor
+ExprMutator = expr_functor.ExprMutator
 
 # helper functions
 var = expr.var
